+++ frontend/src/queries/usePltEventsMetricsQuery.ts
@@ -1,42 +1,63 @@
 import { useQuery, gql } from '@urql/vue'
 import type { Ref } from 'vue'
-import type { MetricsPeriod, PltEventMetrics } from '~/types/generated'
-export type PltEventMetricsQueryResponse = {
-	pltEventMetrics: PltEventMetrics
+import type {
+	MetricsPeriod,
+	PltTransferMetrics,
+	PltMetrics,
+} from '~/types/generated'
+
+export type PltMetricsQueryResponse = {
+	pltMetrics: PltMetrics
 }
 
-const PltEventMetricsQuery = gql<PltEventMetricsQueryResponse>`
-	query PltEventMetrics($period: MetricsPeriod!) {
-		pltEventMetrics(period: $period) {
+const PltMetricsQuery = gql<PltMetricsQueryResponse>`
+	query PltMetrics($period: MetricsPeriod!) {
+		pltMetrics(period: $period) {
+			transactionCount
+			transferVolume
+			uniqueAccounts
+		}
+	}
+`
+
+export const usePltMetricsQuery = (period: Ref<MetricsPeriod>) => {
+	const { data, executeQuery, fetching } = useQuery({
+		query: PltMetricsQuery,
+		requestPolicy: 'cache-and-network',
+		variables: { period },
+	})
+
+	return { data, executeQuery, loading: fetching }
+}
+
+export type PltTransferMetricsQueryResponse = {
+	pltTransferMetrics: PltTransferMetrics
+}
+
+const PltTransferMetricsQueryByTokenId = gql<PltTransferMetricsQueryResponse>`
+	query PltTransferMetrics($period: MetricsPeriod!, $tokenId: String!) {
+		pltTransferMetrics(period: $period, tokenId: $tokenId) {
 			transferCount
 			transferVolume
-			mintCount
-			mintVolume
-			burnCount
-			burnVolume
-			tokenModuleCount
-			totalEventCount
+			decimal
 			buckets {
 				bucketWidth
 				x_Time
 				y_TransferCount
 				y_TransferVolume
-				y_MintCount
-				y_MintVolume
-				y_BurnCount
-				y_BurnVolume
-				y_TokenModuleCount
-				y_TotalEventCount
 			}
 		}
 	}
 `
 
-export const usePltEventsMetricsQuery = (period: Ref<MetricsPeriod>) => {
+export const usePltTransferMetricsQueryByTokenId = (
+	period: Ref<MetricsPeriod>,
+	tokenId: Ref<string> | string
+) => {
 	const { data, executeQuery, fetching } = useQuery({
-		query: PltEventMetricsQuery,
+		query: PltTransferMetricsQueryByTokenId,
 		requestPolicy: 'cache-and-network',
-		variables: { period },
+		variables: { period, tokenId },
 	})
 
 	return { data, executeQuery, loading: fetching }